cli: move the generator resolution to lib/resolver and the describe formatting to the cli
`lib/resolver/generator-dependencies.ts` resolves the generators contributing options to a command (imports and
blueprint overrides), `lib/resolver/generator-commands.ts` looks up the generators of the installation with their
command definition and USAGE. `lib/command/describe-command.ts` keeps the description model, the table and prompt
formatting is cli presentation and lives in `cli/describe.ts`.
